Split out upgrade function
