+++ dcos-launch.spec
@@ -6,8 +6,8 @@
 #     with the final exe. Ensure build system has OpenSSL 1.0.2g or greater
 a = Analysis(['dcos_launch/cli.py'],
              hiddenimports=['html.parser'],
-             datas=[('dcos_launch/ip-detect/aws.sh', 'dcos_launch/ip-detect'),
-                    ('dcos_launch/ip-detect/aws_public.sh', 'dcos_launch/ip-detect'),
+             datas=[('dcos_test_utils/ip-detect/aws.sh', 'dcos_test_utils/ip-detect'),
+                    ('dcos_test_utils/ip-detect/aws_public.sh', 'dcos_test_utils/ip-detect'),
                     ('dcos_test_utils/templates/*.json','dcos_test_utils/templates')
 ])
 pyz = PYZ(a.pure, a.zipped_data, cipher=None)
