+++ dcos_test_utils/ip-detect/aws.sh
