+++ dcos_test_utils/ip-detect/aws_public.sh
