+++ tox.ini
@@ -1,3 +1,6 @@
+[tox]
+envlist = py35-{syntax,unit-tests,build-binary}
+
 [flake8]
 max-line-length=120
 exclude=.git,.tox
@@ -18,7 +21,7 @@ deps =
 commands =
   flake8 --verbose
 
-[testenv:py35-unittests]
+[testenv:py35-unit-tests]
 deps = -rrequirements.txt
 commands=
   py.test {posargs}
